Create README.md
Added some quotes to the command for the domains parameter. The shodan api cannot accept comma lists with a space in them and powershell param automatically inserts a space in between them unless the parameter is in quotes. This behavior can be observed by inserting a Write-Host $domains in the same section of the program as $apistring="&key=$api"